COMP: Fix compiler error on clang c++11
When building Slicer on the default VTK repository
the following class of compiler errors needed to
be addressed.

These were a result of using the compiler
/Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/bin/clang++ \

With compiler flags: -stdlib=libc++ -std=c++11

VTK/Filtering/vtkKdTree.cxx:2677:22: error: non-constant-expression
                            cannot be narrowed from type 'double' to
                            'float' in initializer list [-Wc++11-narrowing]
  float xfloat[3] = {x[0], x[1], x[2]};
                     ^~~~

Using explicit static_cast<>() casting avoid the compiler error.

+++ Charts/vtkOpenGL2ContextDevice2D.cxx
@@ -146,10 +146,10 @@ void vtkOpenGL2ContextDevice2D::DrawImage(float p[2], float scale,
   this->SetTexture(image);
   this->Storage->Texture->Render(this->Renderer);
   int *extent = image->GetExtent();
-  float points[] = { p[0]                    , p[1],
-                     p[0]+scale*extent[1]+1.0, p[1],
-                     p[0]+scale*extent[1]+1.0, p[1]+scale*extent[3]+1.0,
-                     p[0]                    , p[1]+scale*extent[3]+1.0 };
+  float points[] = { static_cast<float>(p[0]                    ), static_cast<float>(p[1]),
+                     static_cast<float>(p[0]+scale*extent[1]+1.0), static_cast<float>(p[1]),
+                     static_cast<float>(p[0]+scale*extent[1]+1.0), static_cast<float>(p[1]+scale*extent[3]+1.0),
+                     static_cast<float>(p[0]                    ), static_cast<float>(p[1]+scale*extent[3]+1.0) };
 
   float texCoord[] = { 0.0, 0.0,
                        1.0, 0.0,
