+++ Views/vtkParallelCoordinatesRepresentation.cxx
@@ -381,7 +381,7 @@ void vtkParallelCoordinatesRepresentation::UpdateHoverHighlight(vtkView* view,
     int lineNum = 0;
     vtkIdType* pts = 0;
     vtkIdType  npts = 0;
-    double p[3] = {x,y,0.0};
+    double p[3] = {static_cast<double>(x),static_cast<double>(y),0.0};
     p[0] /= size[0];
     p[1] /= size[1];
